Create nfc_tag_batch_writer.ino
This code is still in early alpha. It has worked as written exactly once but after further testing it seems to consistently hit E104. This code isn't feature complete and anticipates tags to already be NDEF formatted. A future version should probably format unformatted tags and clean already formatted tags if doing so doesn't significantly increase the total write time of each tag. This would eliminate the need for a seperate formating process.