NFC Locker Arduino Uno code and wiring diagram
This NFC Locker code is complete and tested. The wiring diagram is accurate to our test platform with the acception of the solenoid components. Our solenoid has yet to arrive so for testing purposes we utilized an LED in its place. The solenoid wiring is based on the former CSE 453 Alden Middle School Locker project wiring diagram and has not been independently verified by us for accuracy and functionality.